feat(dev): rename file hompeage to homepage
